requirements.txt
Django>=3.0.4,<3.1.0
psycopg2-binary==2.8.4
flake8==3.7.9
Pillow>=7.1.1,<7.2.0
coverage>=5.0.4,<5.1.0
django-allauth>=0.41.0,<0.42.0
django-cropperjs>=0.0.5,<0.1.0
django-markdownx>=3.0.1,<3.1.0
django-filter>=2.2.0,<2.3.0
